Update logrotate
If you use rsyslog - at least on Jessie PI to forward the logfile to some central location rsyslog does not notice the updated logfile. 

After I had added the restart to the post rotate this is working with no issue. As I did not test if this is working on other type of installation that might only be save for a pi based installation.